+++ kinto/core/storage/postgresql/__init__.py
@@ -67,7 +67,7 @@ class Storage(StorageBase):
 
     """  # NOQA
 
-    schema_version = 13
+    schema_version = 14
 
     def __init__(self, client, max_fetch_size, *args, **kwargs):
         super(Storage, self).__init__(*args, **kwargs)